Initial radbot configuration
Created a new incarnation of RadBot with Lita. Using the following
plugins, most set up to be configured via enviroment:

- aww - Show cute pictures.
- catgif - The stuff the internets are made of...cats.
- cleverbot - Make radbot have a conversation.
- down-for-everyone - Check the status of other websites.
- github-pinger - Ping individuals on slack when they are mentioned on
  GitHub.
- hangout - Create and join hangouts.
- imgflip - Memes never get old.
- karma - Track points. So you know you are winning chat.
- wtf - Define things.

I set it up to work with slack when the slack token is set in the
environment, otherwise Lita expects to be run on the command line (for
great testing).